+++ web/src/containers/UnexpectedError.js
@@ -1,4 +1,4 @@
-import { Alert, Button } from '@cmsgov/design-system-core';
+import { Alert, Button } from '@cmsgov/design-system';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';