+++ web/src/containers/admin/CreateAccount.js
@@ -1,4 +1,4 @@
-import { FormLabel, Select, TextField } from '@cmsgov/design-system-core';
+import { FormLabel, Dropdown, TextField } from '@cmsgov/design-system';
 import PropTypes from 'prop-types';
 import React, { Fragment, useMemo, useState } from 'react';
 import { connect } from 'react-redux';
@@ -83,7 +83,7 @@ const CreateUser = ({ createUser, error, roles, working }) => {
         <FormLabel component="label" fieldId="create_account_state">
           State
         </FormLabel>
-        <Select
+        <Dropdown
           id="create_account_state"
           name="state"
           size="medium"
@@ -96,12 +96,12 @@ const CreateUser = ({ createUser, error, roles, working }) => {
               {s.name}
             </option>
           ))}
-        </Select>
+        </Dropdown>
 
         <FormLabel component="label" fieldId="create_account_role">
           Authorization role
         </FormLabel>
-        <Select
+        <Dropdown
           id="create_account_role"
           name="role"
           size="medium"
@@ -114,7 +114,7 @@ const CreateUser = ({ createUser, error, roles, working }) => {
               {toSentenceCase(r.name)}
             </option>
           ))}
-        </Select>
+        </Dropdown>
 
         <Password
           value={password}