+++ web/src/containers/admin/EditAccount.js
@@ -1,9 +1,9 @@
 import {
   Button,
   FormLabel,
-  Select,
+  Dropdown,
   TextField
-} from '@cmsgov/design-system-core';
+} from '@cmsgov/design-system';
 import PropTypes from 'prop-types';
 import React, { Fragment, useState, useMemo } from 'react';
 import { connect } from 'react-redux';
@@ -121,7 +121,7 @@ const EditAccount = ({
           <FormLabel component="label" fieldId="modify_account_state">
             State
           </FormLabel>
-          <Select
+          <Dropdown
             id="modify_account_state"
             name="state"
             size="medium"
@@ -134,12 +134,12 @@ const EditAccount = ({
                 {s.name}
               </option>
             ))}
-          </Select>
+          </Dropdown>
 
           <FormLabel component="label" fieldId="modify_account_role">
             Authorization role
           </FormLabel>
-          <Select
+          <Dropdown
             id="modify_account_role"
             name="role"
             size="medium"
@@ -153,7 +153,7 @@ const EditAccount = ({
                 {toSentenceCase(r.name)}
               </option>
             ))}
-          </Select>
+          </Dropdown>
 
           <div className="ds-l-row ds-u-padding-x--2">
             <TextField
@@ -213,7 +213,7 @@ const EditAccount = ({
         <FormLabel component="label" fieldId="modify_account_user">
           Account to edit
         </FormLabel>
-        <Select
+        <Dropdown
           id="modify_account_user"
           name="userID"
           value={`${userID}`}
@@ -225,7 +225,7 @@ const EditAccount = ({
               {`${u.name ? `${u.name} - ` : ''}${u.username}`}
             </option>
           ))}
-        </Select>
+        </Dropdown>
 
         {getForm()}
       </CardForm>