+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Anticipated release: August 10, 2020
 
 #### 🚀 New features
 
+- Upgrade to @cmsgov/design-system v2.0.0, remove focus styles ([#2412])
+
 #### 🐛 Bugs fixed
 
 #### ⚙️ Behind the scenes
@@ -15,3 +17,4 @@ Anticipated release: August 10, 2020
 See our [release history](https://github.com/18F/cms-hitech-apd/releases)
 
 [#2335]: https://github.com/18F/cms-hitech-apd/issues/2335
+[#2412]: https://github.com/CMSgov/eAPD/issues/2412